+++ src/falconpy/_payload/__init__.py
@@ -112,6 +112,15 @@ from ._d4c_registration import (
     aws_d4c_registration_payload,
     gcp_registration_payload
     )
+from ._data_protection_configuration import(
+    data_protection_classification_payload,
+    data_protection_cloud_app_payload,
+    data_protection_content_pattern_payload,
+    data_protection_enterprise_account_payload,
+    data_protection_sensitivity_label_payload,
+    data_protection_policy_payload,
+    data_protection_web_locations_payload
+    )
 from ._cspm_registration import (
     cspm_registration_payload,
     cspm_policy_payload,
@@ -203,5 +212,8 @@ __all__ = [
     "scheduled_task_payload", "automation_policy_payload", "policy_host_group_payload",
     "automation_live_query_payload", "cloud_policies_rule_update_payload", "cloud_policies_compliance_control_payload",
     "cloud_policies_rule_create_payload", "cloud_policies_rule_override_payload", "cloud_policies_rule_assign_payload",
-    "cloud_policies_evaluation_payload", "cloud_security_create_group_payload", "correlation_rules_admin_payload"
+    "cloud_policies_evaluation_payload", "cloud_security_create_group_payload", "correlation_rules_admin_payload", 
+    "data_protection_classification_payload", "data_protection_cloud_app_payload", "data_protection_content_pattern_payload",
+    "data_protection_enterprise_account_payload", "data_protection_sensitivity_label_payload", "data_protection_policy_payload",
+    "data_protection_web_locations_payload"
 ]